Move fields types to new module
