+++ lib/src/default.rs
@@ -3,14 +3,14 @@
 // You may obtain a copy of the License in the LICENSE-APACHE file or at:
 //     https://www.apache.org/licenses/LICENSE-2.0
 
+use crate::fields::{Fields, FieldsNamed, FieldsUnnamed};
 use crate::generics::{clause_to_toks, WhereClause};
 use crate::{Scope, ScopeAttr, ScopeItem};
 use proc_macro2::{Span, TokenStream};
 use proc_macro_error::emit_error;
-use quote::{quote, ToTokens, TokenStreamExt};
+use quote::quote;
 use syn::parse::{Error, Parse, ParseStream, Result};
-use syn::punctuated::Punctuated;
-use syn::{parse2, token, Attribute, Expr, Generics, Ident, Item, Token, Type, Visibility};
+use syn::{parse2, Expr, Generics, Ident, Item, Token};
 
 /// `#[impl_default]` attribute
 pub struct ImplDefault {
@@ -159,263 +159,5 @@ impl Parse for ImplDefault {
             where_clause,
             span,
         })
-    }
-}
-
-/// Data stored within an enum variant or struct.
-///
-/// This is a variant of [`syn::Fields`] supporting field initializers.
-#[derive(Debug)]
-pub enum Fields {
-    /// Named fields of a struct or struct variant such as `Point { x: f64, y: f64 }`.
-    Named(FieldsNamed),
-    /// Unnamed fields of a tuple struct or tuple variant such as `Some(T)`.
-    Unnamed(FieldsUnnamed),
-    /// Unit struct or unit variant such as `None`.
-    Unit,
-}
-
-/// Named fields of a struct or struct variant such as `Point { x: f64, y: f64 }`.
-///
-/// This is a variant of [`syn::FieldsNamed`] supporting field initializers.
-#[derive(Debug)]
-pub struct FieldsNamed {
-    /// `{ ... }` around fields
-    pub brace_token: token::Brace,
-    /// Fields
-    pub fields: Punctuated<Field, Token![,]>,
-}
-
-/// Unnamed fields of a tuple struct or tuple variant such as `Some(T)`.
-///
-/// This is a variant of [`syn::FieldsUnnamed`] supporting field initializers.
-#[derive(Debug)]
-pub struct FieldsUnnamed {
-    /// `( ... )` around fields
-    pub paren_token: token::Paren,
-    /// Fields
-    pub fields: Punctuated<Field, Token![,]>,
-}
-
-/// A field of a struct or enum variant.
-///
-/// This is a variant of [`syn::Field`] supporting field initializers.
-#[derive(Debug)]
-pub struct Field {
-    /// Attributes tagged on the field.
-    pub attrs: Vec<Attribute>,
-    /// Visibility of the field.
-    pub vis: Visibility,
-    /// Name of the field, if any.
-    pub ident: Option<Ident>,
-    /// `:` token before type
-    pub colon_token: Option<Token![:]>,
-    /// Type of the field.
-    pub ty: Type,
-    /// Optional field initializer.
-    ///
-    /// This is considered legal input when parsing, but not legal output. An
-    /// attribute rule such as [`ATTR_IMPL_DEFAULT`] must remove the initializer
-    /// before output is generated.
-    pub assign: Option<(Token![=], Expr)>,
-}
-
-// Copied from syn, modified
-pub(crate) mod parsing {
-    use super::*;
-    use syn::ext::IdentExt;
-    use syn::{braced, parenthesized, WhereClause};
-
-    impl Parse for FieldsNamed {
-        fn parse(input: ParseStream) -> Result<Self> {
-            let content;
-            let brace_token = braced!(content in input);
-            Ok(FieldsNamed {
-                brace_token,
-                fields: content.parse_terminated(Field::parse_named)?,
-            })
-        }
-    }
-
-    impl Parse for FieldsUnnamed {
-        fn parse(input: ParseStream) -> Result<Self> {
-            let content;
-            let paren_token = parenthesized!(content in input);
-            Ok(FieldsUnnamed {
-                paren_token,
-                fields: content.parse_terminated(Field::parse_unnamed)?,
-            })
-        }
-    }
-
-    impl Field {
-        /// Parse a named field
-        pub fn parse_named(input: ParseStream) -> Result<Self> {
-            Ok(Field {
-                attrs: input.call(Attribute::parse_outer)?,
-                vis: input.parse()?,
-                ident: Some(if input.peek(Token![_]) {
-                    input.call(Ident::parse_any)
-                } else {
-                    input.parse()
-                }?),
-                colon_token: Some(input.parse()?),
-                ty: input.parse()?,
-                assign: if input.peek(Token![=]) {
-                    Some((input.parse()?, input.parse()?))
-                } else {
-                    None
-                },
-            })
-        }
-
-        /// Parse an unnamed field
-        pub fn parse_unnamed(input: ParseStream) -> Result<Self> {
-            Ok(Field {
-                attrs: input.call(Attribute::parse_outer)?,
-                vis: input.parse()?,
-                ident: None,
-                colon_token: None,
-                ty: input.parse()?,
-                assign: if input.peek(Token![=]) {
-                    Some((input.parse()?, input.parse()?))
-                } else {
-                    None
-                },
-            })
-        }
-    }
-
-    pub(crate) fn data_struct(
-        input: ParseStream,
-    ) -> Result<(Option<WhereClause>, Fields, Option<Token![;]>)> {
-        let mut lookahead = input.lookahead1();
-        let mut where_clause = None;
-        if lookahead.peek(Token![where]) {
-            where_clause = Some(input.parse()?);
-            lookahead = input.lookahead1();
-        }
-
-        if where_clause.is_none() && lookahead.peek(token::Paren) {
-            let fields = input.parse()?;
-
-            lookahead = input.lookahead1();
-            if lookahead.peek(Token![where]) {
-                where_clause = Some(input.parse()?);
-                lookahead = input.lookahead1();
-            }
-
-            if lookahead.peek(Token![;]) {
-                let semi = input.parse()?;
-                Ok((where_clause, Fields::Unnamed(fields), Some(semi)))
-            } else {
-                Err(lookahead.error())
-            }
-        } else if lookahead.peek(token::Brace) {
-            let fields = parse_braced(input)?;
-            Ok((where_clause, Fields::Named(fields), None))
-        } else if lookahead.peek(Token![;]) {
-            let semi = input.parse()?;
-            Ok((where_clause, Fields::Unit, Some(semi)))
-        } else {
-            Err(lookahead.error())
-        }
-    }
-
-    fn parse_braced(input: ParseStream) -> Result<FieldsNamed> {
-        let content;
-        let brace_token = braced!(content in input);
-        let fields = content.parse_terminated(Field::parse_named)?;
-        Ok(FieldsNamed {
-            brace_token,
-            fields,
-        })
-    }
-}
-
-mod printing {
-    use super::*;
-    use std::{iter, slice};
-    use syn::AttrStyle;
-
-    pub struct TokensOrDefault<'a, T: 'a>(pub &'a Option<T>);
-
-    impl<'a, T> ToTokens for TokensOrDefault<'a, T>
-    where
-        T: ToTokens + Default,
-    {
-        fn to_tokens(&self, tokens: &mut TokenStream) {
-            match self.0 {
-                Some(t) => t.to_tokens(tokens),
-                None => T::default().to_tokens(tokens),
-            }
-        }
-    }
-
-    pub trait FilterAttrs<'a> {
-        type Ret: Iterator<Item = &'a Attribute>;
-
-        fn outer(self) -> Self::Ret;
-        fn inner(self) -> Self::Ret;
-    }
-
-    impl<'a> FilterAttrs<'a> for &'a [Attribute] {
-        type Ret = iter::Filter<slice::Iter<'a, Attribute>, fn(&&Attribute) -> bool>;
-
-        fn outer(self) -> Self::Ret {
-            fn is_outer(attr: &&Attribute) -> bool {
-                match attr.style {
-                    AttrStyle::Outer => true,
-                    AttrStyle::Inner(_) => false,
-                }
-            }
-            self.iter().filter(is_outer)
-        }
-
-        fn inner(self) -> Self::Ret {
-            fn is_inner(attr: &&Attribute) -> bool {
-                match attr.style {
-                    AttrStyle::Inner(_) => true,
-                    AttrStyle::Outer => false,
-                }
-            }
-            self.iter().filter(is_inner)
-        }
-    }
-
-    impl ToTokens for FieldsNamed {
-        fn to_tokens(&self, tokens: &mut TokenStream) {
-            self.brace_token.surround(tokens, |tokens| {
-                self.fields.to_tokens(tokens);
-            });
-        }
-    }
-
-    impl ToTokens for FieldsUnnamed {
-        fn to_tokens(&self, tokens: &mut TokenStream) {
-            self.paren_token.surround(tokens, |tokens| {
-                self.fields.to_tokens(tokens);
-            });
-        }
-    }
-
-    impl ToTokens for Field {
-        fn to_tokens(&self, tokens: &mut TokenStream) {
-            tokens.append_all(&self.attrs);
-            self.vis.to_tokens(tokens);
-            if let Some(ident) = &self.ident {
-                ident.to_tokens(tokens);
-                TokensOrDefault(&self.colon_token).to_tokens(tokens);
-            }
-            self.ty.to_tokens(tokens);
-
-            if let Some(ref assign) = self.assign {
-                emit_error!(
-                    assign.0,
-                    "default value on `struct` field in output";
-                    help = "did you mean to use the `#[impl_default]` attribute?",
-                );
-            }
-        }
     }
 }
